requirements.txt
anthropic
flask
streamlit
python-dotenv
requests
beautifulsoup4
llama-index-core
llama-index-llms-anthropic
llama-index-embeddings-huggingface
tavily-python